# apps/front/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/gate/terminals": {
      "post": {
        "operationId": "issue_gate_terminal",
        "tags": [
          "gate"
        ],
        "description": "게이트 단말 자격증명 발급 (관리 권한)\n무작위 토큰을 생성하고 SHA-256 해시만 저장한다. 평문은 이 응답으로만 전달된다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueTerminalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueTerminalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
    "/gate/terminals/{terminalid}": {
      "delete": {
        "operationId": "revoke_gate_terminal",
        "tags": [
          "gate"
        ],
        "description": "게이트 단말 폐기 (관리 권한) — 행을 지우지 않고 is_active만 false로 바꾼다 (소프트 삭제)\n이미 폐기된 단말도 멱등하게 200을 반환한다.",
        "parameters": [
          {
            "name": "terminalid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeTerminalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
    "/gate/verify": {
      "post": {
        "operationId": "verify_gate",
        "tags": [
          "gate"
        ],
        "description": "게이트 사원증 태깅 검증 (FR-D5, 데모 시나리오 1)\n게이트 단말 자격증명(Bearer 토큰)으로만 호출 가능하다.\n(지침 확인 + 장비 완료 + 승인) 충족 시 통과 처리, 아니면 사유와 함께 차단.\n통과 판정은 (출근 건, 작업일) 유니크 제약으로 멱등 기록된다(AC-5).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GateVerifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GateVerifyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
...
    "/work-assignments": {
      "post": {
        "operationId": "create_work_assignment",
        "tags": [
          "work_assignments"
        ],
        "description": "작업 배정 생성 (FR-F3, 데모 시나리오 6 — 관리 권한)\n배정일은 요청 당일로 기록하고, 배정 담당자는 호출한 관리자 계정으로 남긴다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkAssignmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkAssignmentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/work-assignments/{id}": {
      "patch": {
        "operationId": "update_work_assignment",
        "tags": [
          "work_assignments"
        ],
        "description": "작업 배정 수정 (관리 권한)\n배정 대상 화물을 변경하는 경우에도 검수 확정 여부를 동일하게 강제한다 (시나리오 6).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkAssignmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkAssignmentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AssignmentErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "기계 판독용 사유 코드 (예: MSDS_REVIEW_NOT_CONFIRMED)"
          },
          "message": {
            "type": "string",
            "description": "작업자에게 표시할 한국어 사유"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "CreateWorkAssignmentRequest": {
        "type": "object",
        "properties": {
          "cargo_item_id": {
            "type": "integer",
            "format": "int64",
            "description": "배정 대상 화물 고유번호 (없으면 null — 공용 작업)",
            "nullable": true
          },
          "employee_id": {
            "type": "integer",
            "format": "int64",
            "description": "작업자 직원 고유번호"
          }
        },
        "required": [
          "employee_id"
        ]
      },
...
      "EligibilityStatus": {
        "type": "string",
        "enum": [
          "ELIGIBLE",
          "EXCLUDED"
        ]
      },
...
      "GateVerifyRequest": {
        "type": "object",
        "properties": {
          "nfc_card_uid": {
            "type": "string",
            "description": "사원증 NFC UID (FR-D5)"
          }
        },
        "required": [
          "nfc_card_uid"
        ]
      },
      "GateVerifyResponse": {
        "type": "object",
        "properties": {
          "allowed": {
            "type": "boolean"
          },
          "employee_name": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "allowed",
          "employee_name",
          "reason"
        ]
      },
...
      "IssueTerminalRequest": {
        "type": "object",
        "properties": {
          "gate_id": {
            "type": "string",
            "description": "게이트 식별자 (예: 북문, 남문)"
          }
        },
        "required": [
          "gate_id"
        ]
      },
      "IssueTerminalResponse": {
        "type": "object",
        "properties": {
          "gate_id": {
            "type": "string"
          },
          "terminal_id": {
            "type": "integer",
            "format": "int64",
            "description": "단말 고유번호"
          },
          "token": {
            "type": "string",
            "description": "단말 토큰 평문 — 발급 시 단 한 번만 반환되며 다시 조회할 수 없다"
          }
        },
        "required": [
          "terminal_id",
          "gate_id",
          "token"
        ]
      },
...
      "RevokeTerminalResponse": {
        "type": "object",
        "properties": {
          "gate_id": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "terminal_id": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "terminal_id",
          "gate_id",
          "is_active"
        ]
      },
...
      "UpdateWorkAssignmentRequest": {
        "type": "object",
        "properties": {
          "cargo_item_id": {
            "type": "integer",
            "format": "int64",
            "description": "변경할 배정 대상 화물 고유번호 (생략 시 기존 값 유지)",
            "nullable": true
          },
          "eligibility_status": {
            "$ref": "#/components/schemas/EligibilityStatus",
            "description": "ELIGIBLE | EXCLUDED (생략 시 기존 값 유지)",
            "nullable": true
          }
        }
      },
...
      "WorkAssignmentResponse": {
        "type": "object",
        "properties": {
          "assigned_by_id": {
            "type": "integer",
            "format": "int64"
          },
          "assignment_id": {
            "type": "integer",
            "format": "int64"
          },
          "cargo_item_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          },
          "eligibility_status": {
            "type": "string",
            "description": "ELIGIBLE | EXCLUDED"
          },
          "employee_id": {
            "type": "integer",
            "format": "int64"
          },
          "work_date": {
            "type": "string"
          }
        },
        "required": [
          "assignment_id",
          "employee_id",
          "work_date",
          "assigned_by_id",
          "eligibility_status",
          "created_at"
        ]
      },
...
    }
  },
...
}